requirements.txt
numpy
flask
Pillow
sympy
Flask
adafruit-circuitpython-neopixel
rpi_ws281x
opencv-python
